Add TODO to channel for alternative strategy.
This strategy would involve keeping a shared_ptr to the SPModelSpace in each channel and not storing the channel. Instead the channel key (already stored in the channel) can be used to obtain the channel for each state via the pointer and a lookup.

A more severe strategy would involve compressing the channel key using a wrapped index of sorts for the single-particle channels and only storing the additional operator channels.